+++ src/no_drop_msg.rs
@@ -6,11 +6,24 @@ use std::ptr;
 /// [`Self::consume`]d or [`Self::forget`].
 ///
 /// The lifetime parameter `'msg` allows borrowing the message, and most commonly will be `'static`.
-#[derive(Debug, derive_more::Deref, derive_more::DerefMut)]
+#[derive(
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Debug,
+    derive_more::Deref,
+    derive_more::DerefMut,
+    derive_more::AsMut,
+    derive_more::AsRef,
+)]
 #[must_use]
 pub struct NoDropMsg<'msg, T = ()> {
     #[deref]
     #[deref_mut]
+    #[as_mut]
+    #[as_ref]
     value: T,
     msg: Cow<'msg, str>,
 }
@@ -23,10 +36,11 @@ impl<'msg, T> NoDropMsg<'msg, T> {
     /// ```rust
     /// use no_drop::rls::NoDropMsg;
     ///
-    /// let wrapper = NoDropMsg::wrap_msg(42, "forgot to process this value");
+    /// let wrapper = NoDropMsg::wrap(42, "forgot to process this value");
     /// assert_eq!(wrapper.consume(), 42);
     /// ```
-    pub fn wrap_msg<M: Into<Cow<'msg, str>>>(value: T, msg: M) -> Self {
+    #[inline]
+    pub fn wrap<M: Into<Cow<'msg, str>>>(value: T, msg: M) -> Self {
         Self { value, msg: msg.into() }
     }
 
@@ -42,6 +56,22 @@ impl<'msg, T> NoDropMsg<'msg, T> {
     #[inline]
     pub fn forget(self) {
         let _ = ManuallyDrop::new(self);
+    }
+}
+
+impl<'msg> NoDropMsg<'msg, ()> {
+    /// Creates a new empty [`NoDropMsg`] value with a custom panic `msg`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use no_drop::rls::NoDropMsg;
+    ///
+    /// let wrapper = NoDropMsg::expect("this should be consumed");
+    /// wrapper.forget();
+    /// ```
+    pub fn expect<M: Into<Cow<'msg, str>>>(msg: M) -> Self {
+        Self { value: (), msg: msg.into() }
     }
 }
 
@@ -70,7 +100,7 @@ pub struct NoDropMsgPassthrough<'msg, T = ()> {
 #[allow(dead_code)]
 impl<'msg, T> NoDropMsgPassthrough<'msg, T> {
     /// Creates a new wrapper around `value` with a message. The message is ignored.
-    pub fn wrap_msg<M: Into<Cow<'msg, str>>>(value: T, _msg: M) -> Self {
+    pub fn wrap<M: Into<Cow<'msg, str>>>(value: T, _msg: M) -> Self {
         Self { value, _lifetime: std::marker::PhantomData }
     }
 
@@ -87,6 +117,14 @@ impl<'msg, T> NoDropMsgPassthrough<'msg, T> {
     }
 }
 
+#[allow(dead_code)]
+impl<'msg> NoDropMsgPassthrough<'msg, ()> {
+    /// Creates a new empty `NoDropMsgPassthrough` value. The message is ignored.
+    pub fn expect<M: Into<Cow<'msg, str>>>(_msg: M) -> Self {
+        Self { value: (), _lifetime: std::marker::PhantomData }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -96,24 +134,35 @@ mod tests {
     #[test]
     #[should_panic(expected = "custom panic message")]
     fn no_drop_msg_panics() {
-        let wrapper = NoDropMsg::wrap_msg(42, "custom panic message");
+        let wrapper = NoDropMsg::wrap(42, "custom panic message");
         drop(wrapper);
     }
 
-    test_ctor!(no_drop_msg_static_str, NoDropMsg::wrap_msg, (42, "custom message"), 42);
-    test_ctor!(no_drop_msg_string, NoDropMsg::wrap_msg, (42, String::from("owned message")), 42);
-    test_ctor!(no_drop_msg_passthrough, NoDropMsgPassthrough::wrap_msg, (42, "message"), 42);
+    test_ctor!(no_drop_msg_static_str, NoDropMsg::wrap, (42, "custom message"), 42);
+    test_ctor!(no_drop_msg_string, NoDropMsg::wrap, (42, String::from("owned message")), 42);
+    test_ctor!(no_drop_msg_passthrough, NoDropMsgPassthrough::wrap, (42, "message"), 42);
 
     test_ctor!(into_no_drop_msg_dbg_trait, IntoNoDropDbg::no_drop_msg, (42, "msg"), 42);
     test_ctor!(into_no_drop_msg_rls_trait, IntoNoDropRls::no_drop_msg, (42, "msg"), 42);
 
-    test_forget!(no_drop_msg_forget, NoDropMsg::wrap_msg, (42, "custom message"));
-    test_forget!(no_drop_msg_passthrough_forget, NoDropMsgPassthrough::wrap_msg, (42, "custom message"));
+    test_ctor!(no_drop_msg_expect_static_str, NoDropMsg::expect, ("expected message"), ());
+    test_ctor!(no_drop_msg_expect_string, NoDropMsg::expect, (String::from("owned expected message")), ());
+    test_ctor!(no_drop_msg_passthrough_expect, NoDropMsgPassthrough::expect, ("expected message"), ());
+
+    test_forget!(no_drop_msg_forget, NoDropMsg::wrap, (42, "custom message"));
+    test_forget!(no_drop_msg_passthrough_forget, NoDropMsgPassthrough::wrap, (42, "custom message"));
 
     #[test]
     fn no_drop_msg_borrowed() {
         let msg = String::from("borrowed message");
-        let wrapper = NoDropMsg::wrap_msg(42, msg.as_str());
+        let wrapper = NoDropMsg::wrap(42, msg.as_str());
         assert_eq!(wrapper.consume(), 42);
+    }
+
+    #[test]
+    #[should_panic(expected = "unit value must be consumed")]
+    fn no_drop_msg_expect_panics() {
+        let wrapper = NoDropMsg::expect("unit value must be consumed");
+        drop(wrapper);
     }
 }